+++ app/gui2/src/components/visualizations/TableVisualization.vue
@@ -1,35 +1,15 @@
 <script lang="ts">
 import icons from '@/assets/icons.svg'
-import {
-  clipboardNodeData,
-  tsvTableToEnsoExpression,
-  writeClipboard,
-} from '@/components/GraphEditor/clipboard'
 import TextFormattingSelector from '@/components/TextFormattingSelector.vue'
+import AgGridTableView from '@/components/widgets/AgGridTableView.vue'
 import { Ast } from '@/util/ast'
 import { Pattern } from '@/util/ast/match'
-import { useAutoBlur } from '@/util/autoBlur'
 import { VisualizationContainer, useVisualizationConfig } from '@/util/visualizationBuiltins'
 import '@ag-grid-community/styles/ag-grid.css'
 import '@ag-grid-community/styles/ag-theme-alpine.css'
-import type {
-  CellClassParams,
-  CellClickedEvent,
-  ColumnResizedEvent,
-  ICellRendererParams,
-  RowHeightParams,
-} from 'ag-grid-community'
-import type { ColDef, GridOptions } from 'ag-grid-enterprise'
-import {
-  computed,
-  onMounted,
-  onUnmounted,
-  reactive,
-  ref,
-  shallowRef,
-  watchEffect,
-  type Ref,
-} from 'vue'
+import type { CellClassParams, CellClickedEvent, ICellRendererParams } from 'ag-grid-community'
+import type { ColDef } from 'ag-grid-enterprise'
+import { computed, onMounted, ref, shallowRef, watchEffect, type Ref } from 'vue'
 
 export const name = 'Table'
 export const icon = 'table'
@@ -101,23 +81,9 @@ export enum TextFormatOptions {
   On,
   Off,
 }
-
-declare module 'ag-grid-enterprise' {
-  // These type parameters are defined on the original interface.
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  interface AbstractColDef<TData, TValue> {
-    field: string
-  }
-}
-
-if (typeof import.meta.env.VITE_ENSO_AG_GRID_LICENSE_KEY !== 'string') {
-  console.warn('The AG_GRID_LICENSE_KEY is not defined.')
-}
 </script>
 
 <script setup lang="ts">
-const { LicenseManager, Grid } = await import('ag-grid-enterprise')
-
 const props = defineProps<{ data: Data }>()
 const emit = defineEmits<{
   'update:preprocessor': [module: string, method: string, ...args: string[]]
@@ -135,7 +101,6 @@ const SQLITE_CONNECTIONS_NODE_TYPE =
   'Standard.Database.Internal.SQLite.SQLite_Connection.SQLite_Connection'
 const POSTGRES_CONNECTIONS_NODE_TYPE =
   'Standard.Database.Internal.Postgres.Postgres_Connection.Postgres_Connection'
-const DEFAULT_ROW_HEIGHT = 22
 
 const rowLimit = ref(0)
 const page = ref(0)
@@ -143,11 +108,8 @@ const pageLimit = ref(0)
 const rowCount = ref(0)
 const showRowCount = ref(true)
 const isTruncated = ref(false)
-const tableNode = ref<HTMLElement>()
 const dataGroupingMap = shallowRef<Map<string, boolean>>()
-useAutoBlur(tableNode)
-const widths = reactive(new Map<string, number>())
-const defaultColDef = {
+const defaultColDef: Ref<ColDef> = ref({
   editable: false,
   sortable: true,
   filter: true,
@@ -155,21 +117,10 @@ const defaultColDef = {
   minWidth: 25,
   cellRenderer: cellRenderer,
   cellClass: cellClass,
-}
-const agGridOptions: Ref<GridOptions & Required<Pick<GridOptions, 'defaultColDef'>>> = ref({
-  headerHeight: 26,
-  getRowHeight: getRowHeight,
-  rowData: [],
-  columnDefs: [],
-  defaultColDef: defaultColDef as typeof defaultColDef & { manuallySized: boolean },
-  onFirstDataRendered: updateColumnWidths,
-  onRowDataUpdated: updateColumnWidths,
-  onColumnResized: lockColumnSize,
-  sendToClipboard: ({ data }: { data: string }) => sendToClipboard(data),
-  suppressFieldDotNotation: true,
-  enableRangeSelection: true,
-  popupParent: document.body,
-})
+} satisfies ColDef)
+const rowData = ref<Record<string, any>[]>([])
+const columnDefs: Ref<ColDef[]> = ref([])
+
 const textFormatterSelected = ref<TextFormatOptions>(TextFormatOptions.Partial)
 const updateTextFormat = (option: TextFormatOptions) => {
   textFormatterSelected.value = option
@@ -305,29 +256,6 @@ function escapeHTML(str: string) {
     '>': '&gt;',
   }
   return str.replace(/[&<>"']/g, (m) => mapping[m]!)
-}
-
-function getRowHeight(params: RowHeightParams): number {
-  if (textFormatterSelected.value === TextFormatOptions.Off) {
-    return DEFAULT_ROW_HEIGHT
-  }
-
-  const rowData = Object.values(params.data)
-  const textValues = rowData.filter((r): r is string => typeof r === 'string')
-
-  if (!textValues.length) {
-    return DEFAULT_ROW_HEIGHT
-  }
-
-  const returnCharsCount = textValues.map((text: string) => {
-    const crlfCount = (text.match(/\r\n/g) || []).length
-    const crCount = (text.match(/\r/g) || []).length
-    const lfCount = (text.match(/\n/g) || []).length
-    return crCount + lfCount - crlfCount
-  })
-
-  const maxReturnCharsCount = Math.max(...returnCharsCount)
-  return (maxReturnCharsCount + 1) * DEFAULT_ROW_HEIGHT
 }
 
 function cellClass(params: CellClassParams) {
@@ -484,54 +412,48 @@ watchEffect(() => {
         // eslint-disable-next-line camelcase
         get_child_node: undefined,
       }
-  const options = agGridOptions.value
-  if (options.api == null) {
-    return
-  }
-
-  let columnDefs: ColDef[] = []
-  let rowData: object[] = []
 
   if ('error' in data_) {
-    columnDefs = [
+    columnDefs.value = [
       {
         field: 'Error',
         cellStyle: { 'white-space': 'normal' },
       },
     ]
-    rowData = [{ Error: data_.error }]
+    rowData.value = [{ Error: data_.error }]
   } else if (data_.type === 'Matrix') {
-    columnDefs.push(toLinkField(INDEX_FIELD_NAME, data_.get_child_node))
+    columnDefs.value = [toLinkField(INDEX_FIELD_NAME, data_.get_child_node)]
     for (let i = 0; i < data_.column_count; i++) {
-      columnDefs.push(toField(i.toString()))
+      columnDefs.value.push(toField(i.toString()))
     }
-    rowData = addRowIndex(data_.json)
+    rowData.value = addRowIndex(data_.json)
     isTruncated.value = data_.all_rows_count !== data_.json.length
   } else if (data_.type === 'Object_Matrix') {
-    columnDefs.push(toLinkField(INDEX_FIELD_NAME, data_.get_child_node))
+    columnDefs.value = [toLinkField(INDEX_FIELD_NAME, data_.get_child_node)]
     let keys = new Set<string>()
     for (const val of data_.json) {
       if (val != null) {
         Object.keys(val).forEach((k) => {
           if (!keys.has(k)) {
             keys.add(k)
-            columnDefs.push(toField(k))
+            columnDefs.value.push(toField(k))
           }
         })
       }
     }
-    rowData = addRowIndex(data_.json)
+    rowData.value = addRowIndex(data_.json)
     isTruncated.value = data_.all_rows_count !== data_.json.length
   } else if (data_.type === 'Excel_Workbook') {
-    columnDefs = [toLinkField('Value', data_.get_child_node)]
-    rowData = data_.sheet_names.map((name) => ({ Value: name }))
+    columnDefs.value = [toLinkField('Value', data_.get_child_node)]
+    rowData.value = data_.sheet_names.map((name) => ({ Value: name }))
   } else if (Array.isArray(data_.json)) {
-    columnDefs = [toLinkField(INDEX_FIELD_NAME, data_.get_child_node), toField('Value')]
-    rowData = data_.json.map((row, i) => ({ [INDEX_FIELD_NAME]: i, Value: toRender(row) }))
+    columnDefs.value = [toLinkField(INDEX_FIELD_NAME, data_.get_child_node), toField('Value')]
+    rowData.value = data_.json.map((row, i) => ({ [INDEX_FIELD_NAME]: i, Value: toRender(row) }))
     isTruncated.value = data_.all_rows_count ? data_.all_rows_count !== data_.json.length : false
   } else if (data_.json !== undefined) {
-    columnDefs = data_.links ? [toLinkField('Value', data_.get_child_node)] : [toField('Value')]
-    rowData =
+    columnDefs.value =
+      data_.links ? [toLinkField('Value', data_.get_child_node)] : [toField('Value')]
+    rowData.value =
       data_.links ?
         data_.links.map((link) => ({
           Value: link,
@@ -547,15 +469,15 @@ watchEffect(() => {
         return toField(v, valueType)
       }) ?? []
 
-    columnDefs =
+    columnDefs.value =
       data_.has_index_col ?
         [toLinkField(INDEX_FIELD_NAME, data_.get_child_node), ...dataHeader]
       : dataHeader
     const rows = data_.data && data_.data.length > 0 ? data_.data[0]?.length ?? 0 : 0
-    rowData = Array.from({ length: rows }, (_, i) => {
+    rowData.value = Array.from({ length: rows }, (_, i) => {
       const shift = data_.has_index_col ? 1 : 0
       return Object.fromEntries(
-        columnDefs.map((h, j) => {
+        columnDefs.value.map((h, j) => {
           return [
             h.field,
             toRender(h.field === INDEX_FIELD_NAME ? i : data_.data?.[j - shift]?.[i]),
@@ -563,7 +485,7 @@ watchEffect(() => {
         }),
       )
     })
-    isTruncated.value = data_.all_rows_count !== rowData.length
+    isTruncated.value = data_.all_rows_count !== rowData.value.length
   }
 
   // Update paging
@@ -576,11 +498,11 @@ watchEffect(() => {
     page.value = newPageLimit
   }
 
-  if (rowData.length) {
-    const headers = Object.keys(rowData[0] as object)
+  if (rowData.value[0]) {
+    const headers = Object.keys(rowData.value[0])
     const headerGroupingMap = new Map()
     headers.forEach((header) => {
-      const needsGrouping = rowData.some((row: any) => {
+      const needsGrouping = rowData.value.some((row) => {
         if (header in row && row[header] != null) {
           const value = typeof row[header] === 'object' ? row[header].value : row[header]
           return value > 9999
@@ -591,73 +513,10 @@ watchEffect(() => {
     dataGroupingMap.value = headerGroupingMap
   }
 
-  // If an existing grid, merge width from manually sized columns.
-  const newWidths = new Map<string, number>()
-  const mergedColumnDefs = columnDefs.map((columnDef) => {
-    if (!columnDef.field) return columnDef
-    const width = widths.get(columnDef.field)
-    if (width != null) newWidths.set(columnDef.field, (columnDef.width = width))
-    return columnDef
-  })
-  widths.clear()
-  for (const [key, value] of newWidths) widths.set(key, value)
-
   // If data is truncated, we cannot rely on sorting/filtering so will disable.
-  options.defaultColDef.filter = !isTruncated.value
-  options.defaultColDef.sortable = !isTruncated.value
-  options.api.setColumnDefs(mergedColumnDefs)
-  options.api.setRowData(rowData)
+  defaultColDef.value.filter = !isTruncated.value
+  defaultColDef.value.sortable = !isTruncated.value
 })
-
-function updateColumnWidths() {
-  const columnApi = agGridOptions.value.columnApi
-  if (columnApi == null) {
-    console.warn('AG Grid column API does not exist.')
-    return
-  }
-  const cols = columnApi.getAllGridColumns().filter((c) => {
-    const field = c.getColDef().field
-    return field && !widths.has(field)
-  })
-  columnApi.autoSizeColumns(cols)
-}
-
-function lockColumnSize(e: ColumnResizedEvent) {
-  // Check if the resize is finished, and it's not from the API (which is triggered by us).
-  if (!e.finished || e.source === 'api') return
-  // If the user manually resized (or manually autosized) a column, we don't want to auto-size it
-  // on a resize.
-  const manuallySized = e.source !== 'autosizeColumns'
-  for (const column of e.columns ?? []) {
-    const field = column.getColDef().field
-    if (field && manuallySized) widths.set(field, column.getActualWidth())
-  }
-}
-
-/** Copy the provided TSV-formatted table data to the clipboard.
- *
- * The data will be copied as `text/plain` TSV data for spreadsheet applications, and an Enso-specific MIME section for
- * pasting as a new table node.
- *
- * By default, AG Grid writes only `text/plain` TSV data to the clipboard. This is sufficient to paste into spreadsheet
- * applications, which are liberal in what they try to interpret as tabular data; however, when pasting into Enso, the
- * application needs to be able to distinguish tabular clipboard contents to choose the correct paste action.
- *
- * Our heuristic to identify clipboard data from applications like Excel and Google Sheets is to check for a <table> tag
- * in the clipboard `text/html` data. If we were to add a `text/html` section to the data so that it could be recognized
- * like other spreadsheets, when pasting into other applications some applications might use the `text/html` data in
- * preference to the `text/plain` content--so we would need to construct an HTML table that fully represents the
- * content.
- *
- * To avoid that complexity, we bypass our table-data detection by including application-specific data in the clipboard
- * content. This data contains a ready-to-paste node that constructs an Enso table from the provided TSV.
- */
-function sendToClipboard(tsvData: string) {
-  return writeClipboard({
-    ...clipboardNodeData([{ expression: tsvTableToEnsoExpression(tsvData) }]),
-    'text/plain': tsvData,
-  })
-}
 
 // ===============
 // === Updates ===
@@ -665,33 +524,6 @@ function sendToClipboard(tsvData: string) {
 
 onMounted(() => {
   setRowLimit(1000)
-  const agGridLicenseKey = import.meta.env.VITE_ENSO_AG_GRID_LICENSE_KEY
-  if (typeof agGridLicenseKey === 'string') {
-    LicenseManager.setLicenseKey(agGridLicenseKey)
-  } else if (import.meta.env.DEV) {
-    // Hide annoying license validation errors in dev mode when the license is not defined. The
-    // missing define warning is still displayed to not forget about it, but it isn't as obnoxious.
-    const origValidateLicense = LicenseManager.prototype.validateLicense
-    LicenseManager.prototype.validateLicense = function (this) {
-      if (!('licenseManager' in this))
-        Object.defineProperty(this, 'licenseManager', {
-          configurable: true,
-          set(value: any) {
-            Object.getPrototypeOf(value).validateLicense = () => {}
-            delete this.licenseManager
-            this.licenseManager = value
-          },
-        })
-      origValidateLicense.call(this)
-    }
-  }
-  // TODO: consider using Vue component instead: https://ag-grid.com/vue-data-grid/getting-started/
-  new Grid(tableNode.value!, agGridOptions.value)
-  updateColumnWidths()
-})
-
-onUnmounted(() => {
-  agGridOptions.value.api?.destroy()
 })
 </script>
 
@@ -723,7 +555,16 @@ onUnmounted(() => {
           <span v-else v-text="`${rowCount} rows.`"></span>
         </template>
       </div>
-      <div ref="tableNode" class="scrollable ag-theme-alpine"></div>
+      <!-- TODO[ao]: Suspence in theory is not needed here (the entire visualization is inside
+       suspense), but for some reason it causes reactivity loop - see https://github.com/enso-org/enso/issues/10782 -->
+      <Suspense>
+        <AgGridTableView
+          class="scrollable grid"
+          :columnDefs="columnDefs"
+          :rowData="rowData"
+          :defaultColDef="defaultColDef"
+        />
+      </Suspense>
     </div>
   </VisualizationContainer>
 </template>
@@ -736,11 +577,8 @@ onUnmounted(() => {
   height: 100%;
 }
 
-.ag-theme-alpine {
-  --ag-grid-size: 3px;
-  --ag-list-item-height: 20px;
+.grid {
   flex-grow: 1;
-  font-family: var(--font-mono);
 }
 
 .table-visualization-status-bar {